RiskManagement.py
from riskmgmt import RiskManager, VolatilityRiskModel
from interfaces import Binance
from notifiers import send_alert
from cache import RedisCache
import redis

class RedisCache:
    def __init__(self, host='localhost', port=6379):
        self.client = redis.Redis(host=host, port=port)

    def set(self, key, value, ttl=None):
        self.client.set(key, value, ex=ttl)

    def get(self, key):
        return self.client.get(key)

    def memoize(self, func):
        def wrapper(*args, **kwargs):
            key = f"{func.__name__}:{str(args)}:{str(kwargs)}"
            result = self.get(key)
            if result is None:
                result = func(*args, **kwargs)
                self.set(key, result)
            return result
        return wrapper


class RiskManager:
  def __init__(self):
    self.max_sizes = {}
    self.var_limits = {}

  def check_order(self, order):
    # Check if the order meets the risk criteria
    symbol = order['symbol']
    size = order['size']
    return size <= self.max_sizes.get(symbol, float('inf'))

  def set_max_size(self, symbol, alloc):
    # Set the maximum size for a given symbol
    self.max_sizes[symbol] = alloc

  def set_var_limits(self, analysis):
    # Set Value at Risk limits
    self.var_limits = analysis

class VolatilityRiskModel:
    def __init__(self, symbols):
        self.symbols = symbols

    def analyze(self):
        # Perform volatility risk analysis
        pass

    def get_volatility_scores(self):
        # Get volatility scores
        return {}

cache = RedisCache()
rm = RiskManager()
client = Binance()
symbols = ['BTCUSDT', 'ETHUSDT']

def execute_order(order):
    if not rm.check_order(order):
        send_alert("Risk limit breached!")
        return
    client.execute_order(order)

def optimal_allocation(symbol):
    # Calculation for optimal allocation - replace with your logic
    calculated_allocation = 0.1
    return calculated_allocation

for symbol in symbols:
    alloc = optimal_allocation(symbol)
    rm.set_max_size(symbol, alloc)

# Portfolio volatility risk
vol_model = VolatilityRiskModel(symbols)
rm.set_var_limits(vol_model.analyze())


def reduce_leverage(self, vol_scores):
  # Define maximum and minimum leverage levels
  max_leverage = 5
  min_leverage = 1

  # Define thresholds for volatility scores
  high_vol_threshold = 0.6
  low_vol_threshold = 0.3

  # Iterate through the vol_scores and adjust leverage accordingly
  for symbol, vol_score in vol_scores.items():
    # High volatility: Reduce leverage
    if vol_score > high_vol_threshold:
      leverage = min_leverage
    # Low volatility: Increase leverage
    elif vol_score < low_vol_threshold:
      leverage = max_leverage
    # Moderate volatility: Scale leverage linearly
    else:
      leverage = (max_leverage - min_leverage) * (vol_score - low_vol_threshold) / (
                high_vol_threshold - low_vol_threshold) + min_leverage

    # Apply the leverage to the trading strategy
    self.apply_leverage(symbol, leverage)

    # Log or take other actions
    print(f"Set leverage for {symbol}: {leverage}")


def apply_leverage(self, symbol, leverage):
  # Logic to apply the leverage to the trading strategy
  # Update the positions, margin requirements, etc.
  pass

vol_scores = vol_model.get_volatility_scores()
reduce_leverage(vol_scores)

# Caching with RedisCache
@cache.memoize
def optimal_allocation(symbol):
    # Calculation logic here
    calculated_allocation = 0.1
    return calculated_allocation
