requirements.txt
# requirements.txt

python-dotenv~=0.21.1
PythonInterfaces~=0.0.2
notifiers~=1.3.3
cache~=1.0.3
aiohttp~=3.8.5
cryptocmd~=0.6.3
monitoring~=0.9.12
utils~=1.0.1
exchanges~=0.0.1
analytics~=0.6.5
pyspark~=3.4.1
Backtesting~=0.3.3
strategies~=0.2.3
Risk~=0.1.4
django-reporting~=0.21
tuning~=2020.8.28.15.39.17
cachetools==5.3.1
alerting~=0.0.9.3
caching~=0.1.dev8
services~=0.1.1
config~=0.5.1